requirements.txt
bootstrap4==0.1.0
Django==2.2.1
django-allauth==0.40.0
django-crispy-forms==1.7.2
django-filter==2.2.0
openpyxl==3.1.1
Pillow
psycopg2-binary
python-decouple==3.7
qrcode==7.4.2
six==1.12.0
WeasyPrint==50
dj-database-url
django-widget-tweaks
whitenoise
gunicorn
